smp: Add forward tree protocol
Introduce the Forward Tree protocol (FT). The FT route data downstream
in a tree of MCUmgr devices. The Flag 0x80 was selected to indicate
that the data payload contains aditionally the FT 8 bytes. This 8 bytes
are placed at end of regular group payload.

The FT payload is composed by 16 nimbles. The first nimble is the number
of hops down in the tree, which in total can handle 15. The remaining
nimbles identify the downstream port index.

The FT is a generic downstream route protocol mechanism and it is not
designed to create network were a downstream device can send
communication upstream or downstream.

FT example:
                          host
                           |
                           D1
                          /|\
                     ----- | -----
                    /      D3     \
                   D2      |      D4
                          / \
                  D5------   ------D6

The host is the controller which sendis SMP requests. This means that
data flows from host to devices in the downstream direction. A response
from a device to host is the upstream direction.

In this scenario, all upstream messages are processed as a transparent
forward response to the host or without any changes the content.

A downstream message needs to be inspected when the FORWARD_TREE flag
bit (0x80) is set. The FT payload format is compose by 16 nimbles. The
most left nimble is the hops counter. All the other nimbles are port
indexes.

       ---- Nimble 15       ---- Nimble 1
      /                    /
  0xH0'00'00'00'00'00'00'0N
    \
     --- hops

Assuming that the host wants to send data to D4 device:

1- set the hops property to 1
2- add the FT port[0] nimble as 2

The correspondent FT big endian data payload in the wire will be:

                            ---- Nimble 1
                           /
  0x10'00'00'00'00'00'00'02
    \
     --- hops

In this case, when host send the data to D1 the device will inspect the
content and forward downstream in the D1 port[2], where D4 is connected.

In the same way, when host wants to send a request to D6:

1- set the hops property to 2
2- set the FT port[1] nimble to 1 and port[0] nimble to 1

                            ---- Ninble 1
                           /---- Nimble 0
                          //
  0x20'00'00'00'00'00'00'11
    \
     --- hops

This means that each time a device must inspect the FT payload and
forward downstream the hops count should be reduced by 1 and when hops
is evaluated to 0 the device must consume the package.

The system will use the hops value to index the correspondent FT port
nimble. In summary, the FT lower port index represents the route closer
to the destine and the high port index represents the node closer to
the device that start the request.

Signed-off-by: Gerson Fernando Budke <gerson.budke@ossystems.com.br>
(cherry picked from commit 08c0e014d276d56f745b869d41555e0a38db6b6d)